마이페이지에서 새로고침시 이메일 안뜨는 문제 해결

+++ gagunokuga/gagunokuga_front/src/views/mypage/mypage/MypageInfo.vue
@@ -13,17 +13,30 @@
 </template>
 
 <script setup>
-import { computed } from "vue";
+import { computed, onMounted  } from "vue";
 import { useLoginStore } from "@/views/login/login";
 import { useRouter } from "vue-router";
 import axios from "axios";
 
 const loginStore = useLoginStore();
+const router = useRouter();
+const baseURL = import.meta.env.VITE_API_URL;
+
 const nickname = computed(() => loginStore.state.nickname);
 const email = computed(() => loginStore.state.email);
 
-const router = useRouter();
-const baseURL = import.meta.env.VITE_API_URL;
+
+// // 컴포넌트 마운트 시 사용자 정보 가져오기
+// onMounted(async () => {
+//   // 토큰이 있는지 확인
+//   const token = localStorage.getItem("accessToken");
+//   if (token) {
+//     loginStore.state.token = token;
+//     await loginStore.fetchUserInfo();
+//   } else {
+//     router.push('/login');
+//   }
+// });
 
 // 프로필 수정 페이지로 이동
 const updateProfileImage = async (file) => {
@@ -38,6 +51,7 @@ const updateProfileImage = async (file) => {
       },
     });
     alert("프로필 이미지가 성공적으로 수정되었습니다.");
+    // await loginStore.fetchUserInfo();
     loginStore.setUserInfo(response.data); // 사용자 정보 업데이트
   } catch (error) {
     console.error("프로필 이미지 수정 오류", error);
@@ -51,13 +65,18 @@ const goToInfoEdit = () => {
 
 // 탈퇴 처리
 const deleteUser = async () => {
+  const isConfirmed = confirm("정말 탈퇴하시겠습니까? ");
+  if (!isConfirmed) {
+    return; 
+  }
+
   try {
     await axios.delete(`${baseURL}/api/users`, {
       headers: {
         Authorization: `Bearer ${loginStore.state.token}`,
       },
     });
-    alert("계정이 삭제되었습니다.");
+    alert("탈퇴되었습니다.");
     loginStore.logout();
     await router.push("/login");
   } catch (error) {
